requirements.txt
torch>=1.10.0
horovod>=0.23.0
deepspeed>=0.6.0
numpy>=1.21.0
pyyaml>=6.0.0